Experiment with a flat layout.
This in a conversion of the belsunu.atf example from
the pyoracc tiny_corpus, placed in a cts file hierarchy
as urn:cts:cdli:X001001.belsunu-akk1. This is flattened
from the typical hierachy in that there's no level 2,
and the textgroup and work urns are the same.

Thomas Köntges and Willis Monroe suggested just doubling
the PNUM e.g. urn:cts:cdli:X001001.X001001.belsunu-akk1,
but that's double the number of files and directories,
so I wanted to see if the tools would accept this instead.