Merge output from recent atf2tei changes.
Add a third level of <div> to account for the different artifacts
under and entry. There's almost always only one, a tablet, but
there could be others, like an envelope.

urns still need to be added manually.  It's not clear we need them
on both text and ti:edition; nemo can load them with just this.