Add cdli catalogue data access code.
This is copied from the cdli-search experiment repository and
wraps the cdli_catalogue_nofm.csv file set from the data export
repo for easier access.

The idea is to use these for filtering the atf records included
in the automatic conversion, limiting the collection to a
demonstration set until atf2cts is more mature.