Offer a containerized version of the conversion.
Add a shell script to manage updating the repo.

When invoked as `./run.sh --setup` it clones the input and
output repositories and generates an ssh key if necessary.

This is intended to be invoked within the container.
Invoke as a long-running process with something like:

    docker build -t cdli-cts-update .
    docker run -d cdli-cts-update ./run.sh setup && ./run.sh

In most cases, one will want to bind-mount existing checkouts
into the running container to save space and startup time.

The python:alpine container image doesn't provide enough information
for git to generate a git author, so one is hard-coded if the cdli-cts
repository checkout is not supplied.. This should really be set per
installation, but this gets our container going for now.